Columnize the logs of the receipts
ClickHouse expects the nested structure as columns in the JSON, so they
need to be converted when extracting the data.

Trimmed some of the fields pf the logs, as they can be found in the receipt.

Also reduced the block batch as it seems the max size of the messages is
reached with the newer blocks.